Summation gpu progress 50% reached 5 Hz

+++ include/nps_uw_sensors_gazebo/gazebo_ros_image_sonar.hh
@@ -73,7 +73,7 @@
 
 namespace gazebo
 {
-  typedef std::complex<double> Complex;
+  typedef std::complex<float> Complex;
   typedef std::valarray<Complex> CArray;
   typedef std::valarray<CArray> CArray2D;
